Move o.g.jvm.{internal.StubbedJar=>tasks.api.ApiJar}
This commit moves StubbedJar from its former home in o.g.jvm.internal to
its new home in o.g.vm.tasks.api, and renames it to ApiJar to better
reflect at a high-level *what* it produces as opposed to reflecting
lower-level details about the process of *how* it gets produced.
