+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/package-info.java
@@ -16,7 +16,7 @@
 
 /**
  * Tasks supporting Gradle's "compile avoidance" feature through the generation and use of
- * API classes and API Jar files.
+ * API classes and {@link org.gradle.jvm.tasks.api.ApiJar} files.
  *
  * @since 2.10
  */
